P0-T3: repo-wide lint baseline green under one unified strict config
- ruff format pass (14 files) + config unification: core's weak
  [tool.ruff]/[tool.mypy]/[tool.pytest] blocks removed — everything now
  resolves to the root strict ruleset (empirically confirmed via
  --show-settings); pyrightconfig.json deleted (unwired).
- True cost was 9x the estimate: with --no-cache the unification exposed
  1138 violations (a stale .ruff_cache had been deflating counts 3-10x —
  recorded in the spec; CI must trust fresh runs, local lint should note it).
- Fixed all families to zero by hand except TC001 (typing-only imports):
  ruff's runtime-evaluated-base-classes=pydantic.BaseModel collapsed the
  false-positive class 621->522; the remainder is a RECORDED descope
  (root-ignore with justification; proper TYPE_CHECKING migration filed for
  the Phase 4 debt wave in the spec's Descope record). Includes 125
  (str,Enum)->StrEnum conversions (member-interpolation surface audited),
  62 ClassVar annotations (no Pydantic field demoted), 3 internal renames
  (zero stale references), 18 rule-specific justified noqas, and a TC004
  circular-import shim annotation.
- Note: spec asked for separate format/fix commits; landed as one after the
  escalation redirect — split not recoverable from the mixed tree.
Gates (--no-cache, triple-verified): format-check clean (1189 files),
ruff check clean, mypy clean (670), suite 4089/26/0.
Spec: build/specs/phase-0-foundation/P0-T3 (+ Descope record).

+++ packages/twinklr/core/feature_engineering/transitions_v2/markov.py
@@ -26,7 +26,7 @@ _LONG_MS = 5000
 _Z = 1.96
 
 # Laplace smoothing alpha — applied lazily only over observed transitions per row,
-# not over every possible (n × n) cell.
+# not over every possible (n x n) cell.
 _ALPHA = 1.0
 
 
@@ -341,7 +341,7 @@ class MarkovTransitionModel:
                     if val > 0:
                         tgt = idx_to_tid.get(j, str(j))
                         # Convert probability back to pseudo-count for sparse storage.
-                        model._counts.setdefault(src, {})[tgt] = int(round(val * 1000))
+                        model._counts.setdefault(src, {})[tgt] = round(val * 1000)
                 model._row_counts[src] = row_counts_list[i] if i < len(row_counts_list) else 0
 
             for b, bmat in bucket_matrices.items():
@@ -351,9 +351,7 @@ class MarkovTransitionModel:
                     for j, val in enumerate(row):
                         if val > 0:
                             tgt = idx_to_tid.get(j, str(j))
-                            model._bucket_counts[b].setdefault(src, {})[tgt] = int(
-                                round(val * 1000)
-                            )
+                            model._bucket_counts[b].setdefault(src, {})[tgt] = round(val * 1000)
                     model._bucket_row_counts[b][src] = brc[i] if i < len(brc) else 0
 
             model._fitted = state.get("fitted", False)
